Refactor serialConsole into new module.
